🧹 token 초기값 변경 및 적용, token검증 렌더링시 하도록 변경
