+++ src/components/post/Post.scss
@@ -18,8 +18,8 @@
       border-radius: 10px;
       margin-bottom: 14.5px;
       .img-post {
-        width: 100%;
         display: block;
+        width: 100%;
       }
     }
     .container-btn-post {
