+++ src/context/UserContext.jsx
@@ -1,13 +1,16 @@
 import { createContext, useState } from "react";
+import axios from "axios";
 
 const UserContext = createContext();
 export default UserContext;
+
 const UserContextProvider = ({ children }) => {
-  const [token, setToken] = useState(localStorage.getItem("token"));
+  const [token, setToken] = useState("Bearer " + localStorage.getItem("token"));
   const [myAccountname, setMyAccountname] = useState(
     localStorage.getItem("accountname")
   );
   const [myImage, setMyImage] = useState();
+
   return (
     <UserContext.Provider
       value={{
