+++ src/pages/profileEditPage/ProfileEdit.jsx
@@ -10,7 +10,6 @@ function ProfileEdit() {
   const location = useLocation();
   const navigate = useNavigate();
   const { token, myAccountname, setMyAccountname } = useContext(UserContext);
-  const authToken = "Bearer " + token;
   const [user, setUser] = useState("");
   const [username, setUsername] = useState(location.state.username);
   const [accountname, setAcountname] = useState(location.state.accountname);
@@ -35,7 +34,7 @@ function ProfileEdit() {
         updatedProfile,
         {
           headers: {
-            Authorization: authToken,
+            Authorization: token,
             "Content-Type": "application/json",
           },
         }
