+++ src/pages/profileEditPage/profileEditInfo/ProfileEditInfo.jsx
@@ -47,7 +47,7 @@ function ProfileEditInfo({
             },
           }
         );
-        if (res.data.message === "사용 가능한 계정ID 입니다.") {
+        if (res.data.message === "사용 가능한 ID 입니다.") {
           setResMessageAccountname(res.data.message);
           setIsValidAccountname(true);
         } else if (myAccountname === accountname) {
